# openapi.yml
openapi: 3.1.0
info:
  title: Tournament
  description: |-
    This API is used as a sandbox to discuss the repository pattern.
  version: 0.0.1
externalDocs:
  description: Source code
  url: https://github.com/sylvaindecout/repository-pattern-example
tags:
  - name: tournament
    description: Back-office operations about tournaments
  - name: users
    description: Back-office operations about users
  - name: selection
    description: Operations related to tournament selection
  - name: roster
    description: Operations about managing player rosters
  - name: player
    description: Operations about a specific player
paths:
  /users/{userId}:
    get:
      tags:
        - users
      summary: Find user
      description: Find a user
      operationId: findUser
      parameters:
        - name: userId
          in: path
          description: ID of user
          required: true
          schema:
            $ref: '#/components/schemas/UserId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid parameters supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - users
      summary: Upsert user
      description: Create or update a user
      operationId: upsertUser
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            $ref: '#/components/schemas/UserId'
      requestBody:
        $ref: '#/components/requestBodies/UserUpdateRequest'
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid parameters supplied
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tournaments/{tournamentId}:
    get:
      tags:
        - tournament
      summary: Find tournament
      description: Find a tournament
      operationId: findTournament
      parameters:
        - name: tournamentId
          in: path
          description: ID of tournament
          required: true
          schema:
            $ref: '#/components/schemas/TournamentId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '400':
          description: Invalid ID supplied
        '404':
          description: Tournament not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - tournament
      summary: Upsert tournament
      description: Create or update a tournament
      operationId: upsertTournament
      parameters:
        - name: tournamentId
          in: path
          description: ID of tournament
          required: true
          schema:
            $ref: '#/components/schemas/TournamentId'
      requestBody:
        $ref: '#/components/requestBodies/TournamentUpdateRequest'
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tournaments/{tournamentId}/players:
    get:
      tags:
        - roster
      summary: List players
      description: List players of a tournament
      operationId: listPlayers
      parameters:
        - name: tournamentId
          in: path
          description: ID of tournament
          required: true
          schema:
            $ref: '#/components/schemas/TournamentId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
                uniqueItems: true
        '400':
          description: Invalid ID supplied
        '404':
          description: Tournament not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - roster
      summary: Add a player
      description: Add a player to the tournament
      operationId: addPlayer
      parameters:
        - name: tournamentId
          in: path
          description: ID of tournament
          required: true
          schema:
            $ref: '#/components/schemas/TournamentId'
      requestBody:
        $ref: '#/components/requestBodies/PlayerCreationRequest'
      responses:
        '201':
          description: Successful operation
          headers:
            Location:
              description: Location of created resource
              schema:
                type: string
                format: uri
                examples: [ 'https://example.com/tournaments/019ca99d-6d8f-7a0d-bbf2-4dbace954c07/players/giorno' ]
        '400':
          description: Invalid ID supplied
        '404':
          description: Tournament not found
        '409':
          description: Conflict
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - roster
      summary: Clear players
      description: Clear players of the tournament
      operationId: clearPlayers
      parameters:
        - name: tournamentId
          in: path
          description: ID of tournament
          required: true
          schema:
            $ref: '#/components/schemas/TournamentId'
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tournaments/{tournamentId}/players/{playerId}:
    get:
      tags:
        - player
      summary: Find a player
      description: Find an existing player of the tournament by ID
      operationId: findPlayer
      parameters:
        - name: tournamentId
          in: path
          description: ID of tournament
          required: true
          schema:
            $ref: '#/components/schemas/TournamentId'
        - name: playerId
          in: path
          description: ID of the player
          required: true
          schema:
            $ref: '#/components/schemas/UserId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '400':
          description: Invalid ID supplied
        '404':
          description: Player not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tournaments/{tournamentId}/players/{playerId}/score:
    put:
      tags:
        - player
      summary: Update score
      description: Update the score of a player in the tournament
      operationId: updateScore
      parameters:
        - name: tournamentId
          in: path
          description: ID of tournament
          required: true
          schema:
            $ref: '#/components/schemas/TournamentId'
        - name: playerId
          in: path
          description: ID of the player
          required: true
          schema:
            $ref: '#/components/schemas/UserId'
      requestBody:
        $ref: '#/components/requestBodies/ScoreUpdateRequest'
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Tournament not found
        '409':
          description: Conflict
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    TournamentSearchResult:
      required:
        - id
        - name
        - maxPlayerRosterSize
        - playerRosterSize
      properties:
        id:
          $ref: '#/components/schemas/TournamentId'
        name:
          $ref: '#/components/schemas/TournamentName'
        maxPlayerRosterSize:
          $ref: '#/components/schemas/PlayerRosterSize'
        minimumAge:
          $ref: '#/components/schemas/Age'
        playerRosterSize:
          $ref: '#/components/schemas/PlayerRosterSize'
        averageAge:
          $ref: '#/components/schemas/Age'
    Tournament:
      required:
        - id
        - name
        - maxPlayerRosterSize
      properties:
        id:
          $ref: '#/components/schemas/TournamentId'
          readonly: true
        name:
          $ref: '#/components/schemas/TournamentName'
        maxPlayerRosterSize:
          $ref: '#/components/schemas/PlayerRosterSize'
        minimumAge:
          $ref: '#/components/schemas/Age'
    User:
      required:
        - id
        - preferredNickname
        - dateOfBirth
        - city
      properties:
        id:
          $ref: '#/components/schemas/UserId'
          readonly: true
        preferredNickname:
          $ref: '#/components/schemas/Nickname'
        dateOfBirth:
          $ref: '#/components/schemas/LocalDate'
        city:
          $ref: '#/components/schemas/City'
    Player:
      required:
        - id
        - nickname
        - age
        - city
        - score
      properties:
        id:
          $ref: '#/components/schemas/UserId'
        nickname:
          $ref: '#/components/schemas/Nickname'
        age:
          $ref: '#/components/schemas/Age'
        city:
          $ref: '#/components/schemas/City'
        score:
          $ref: '#/components/schemas/Score'
    UserId:
      type: string
      format: uuid
    TournamentId:
      type: string
      format: uuid
    TournamentName:
      type: string
      examples: [ 'Tournament #1' ]
    PlayerRosterSize:
      type: integer
      examples: [ 50 ]
    Nickname:
      type: string
      pattern: '[A-Z]{1,50}'
      examples: [ 'giorno', 'jotaro', 'jolyne', 'joseph' ]
    LocalDate:
      type: string
      format: date
    Score:
      type: integer
      minimum: 0
      examples: [ 0, 10, 0, 35 ]
    Age:
      type: integer
      minimum: 16
      examples: [ 20, 35, 25, 21 ]
    City:
      type: string
      minLength: 2
      examples: [ 'naples', 'tokyo', 'orlando', 'new york city' ]
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
  requestBodies:
    PlayerCreationRequest:
      description: Player that needs to be added to the tournament
      content:
        application/json:
          schema:
            required:
              - userId
            properties:
              userId:
                $ref: '#/components/schemas/UserId'
      required: true
    ScoreUpdateRequest:
      content:
        application/json:
          schema:
            required:
              - score
            properties:
              score:
                $ref: '#/components/schemas/Score'
      required: true
    TournamentUpdateRequest:
      description: Tournament that needs to be created or updated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Tournament'
    UserUpdateRequest:
      description: User that needs to be created or updated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/User'